+++ scripts/compile-debug.bat
@@ -38,7 +38,7 @@ if not exist dist mkdir dist
 %CXX% %CXXFLAGS% ^
     -I"%LIBGIT2_INC%" -Iinclude ^
     src\autogitpull.cpp src\git_utils.cpp src\tui.cpp src\logger.cpp src\resource_utils.cpp src\system_utils.cpp src\time_utils.cpp src\debug_utils.cpp src\parse_utils.cpp ^
-    src\config_utils.cpp ^
+    src\config_utils.cpp src\options.cpp ^
     "%LIBGIT2_LIB%\libgit2.a" ^
     -lz -lssh2 -lws2_32 -lwinhttp -lole32 -lrpcrt4 -lcrypt32 -lpsapi -lyaml-cpp ^
     %LDFLAGS% ^
